+++ tools/update_lint_rules/CHANGELOG.md
@@ -1,3 +1,17 @@
-## 1.0.0
+# Changelog
+
+Follow the [Keep a Changelog] format.
+
+## 2.0.0 - 2024-08-15
+
+### Changed
+
+- Remove "group" and change to output "categories".
+
+## 1.0.0 - 2023-09-04
 
 - Initial version.
+
+<!-- Links -->
+
+[Keep a Changelog]: https://keepachangelog.com